LayoutTests/fast/js/kde/encode_decode_uri-expected.txt

KDE JS Test

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


FAIL checkWithFunctions(encodeURI,decodeURI) should be true. Was false
FAIL checkWithFunctions(encodeURIComponent,decodeURIComponent) should be true. Was false
PASS successfullyParsed is true

TEST COMPLETE
